apps/platform-api/pyproject.toml
[project]
name = "platform-api"
version = "0.1.2"
description = "Next-generation platform control plane for AITestLab"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.133.1",
    "uvicorn[standard]>=0.41.0",
    "python-dotenv>=1.2.1",
    "pydantic-settings>=2.11.0",
    "httpx>=0.28.1",
    "langgraph-sdk>=0.3.9",
    "sqlalchemy>=2.0.43",
    "alembic>=1.16.5",
    "psycopg[binary]>=3.3.3",
    "redis>=6.4.0",
    "pyjwt[crypto]>=2.10.1",
    "openpyxl>=3.1.5",
]
